Convert LCAO coefs to single precision if needed

+++ gpaw/new/pw/builder.py
@@ -196,7 +196,8 @@ class PWDFTComponentsBuilder(PWFDDFTComponentsBuilder):
                                comm=self.communicators['b'],
                                xp=self.xp)
             mynbands, M = C_nM.dist.shape
-            phit_aJG.multiply(C_nM.to_xp(self.xp), out_nG=psit_nG[:mynbands])
+            phit_aJG.multiply(C_nM.to_xp(self.xp).to_dtype(pw.dtype),
+                              out_nG=psit_nG[:mynbands])
             return psit_nG
 
         lcao_dtype = complex if \
